Use modifier classname for test-name instead of the modifier counter
If there are modifers the tests are named by the test name plus a counter
based on the modifier position. If you change the sorting or add a new
modifier you will get errors because the images have changed but they are
just stored under a new name. To fix that problem the modifier name is
used instead of the counter.